move contents of weirdly-named provable folder into mina-signer
