+++ src/lib/bool.ts
@@ -1,7 +1,7 @@
 import { Snarky } from '../snarky.js';
 import { Field, readVarMessage, withMessage } from './field.js';
 import { FieldVar, FieldConst, FieldType } from './provable-core/fieldvar.js';
-import { Bool as B } from '../provable/field-bigint.js';
+import { Bool as B } from '../mina-signer/src/field-bigint.js';
 import { defineBinable } from '../bindings/lib/binable.js';
 import { NonNegativeInteger } from '../bindings/crypto/non-negative.js';
 import { asProver } from './provable-context.js';
