+++ src/lib/group.ts
@@ -2,7 +2,7 @@ import { Field } from './field.js';
 import { FieldVar } from './provable-core/fieldvar.js';
 import { Scalar } from './scalar.js';
 import { Snarky } from '../snarky.js';
-import { Field as Fp } from '../provable/field-bigint.js';
+import { Field as Fp } from '../mina-signer/src/field-bigint.js';
 import { GroupAffine, Pallas } from '../bindings/crypto/elliptic-curve.js';
 import { Provable } from './provable.js';
 import { Bool } from './bool.js';
