+++ src/lib/scalar.ts
@@ -1,5 +1,5 @@
 import { Snarky } from '../snarky.js';
-import { Scalar as Fq } from '../provable/curve-bigint.js';
+import { Scalar as Fq } from '../mina-signer/src/curve-bigint.js';
 import { Field } from './field.js';
 import { FieldVar, FieldConst } from './provable-core/fieldvar.js';
 import { MlArray } from './ml/base.js';
