+++ src/mina-signer/src/field-bigint.ts
@@ -1,10 +1,10 @@
-import { randomBytes } from '../bindings/crypto/random.js';
-import { Fp, mod } from '../bindings/crypto/finite-field.js';
+import { randomBytes } from '../../bindings/crypto/random.js';
+import { Fp, mod } from '../../bindings/crypto/finite-field.js';
 import {
   BinableBigint,
   HashInput,
   ProvableBigint,
-} from '../bindings/lib/provable-bigint.js';
+} from '../../bindings/lib/provable-bigint.js';
 
 export { Field, Bool, UInt32, UInt64, Sign };
 export { pseudoClass, sizeInBits, checkRange, checkField };
