+++ src/mina-signer/src/nullifier.ts
@@ -1,12 +1,7 @@
 import { Fq } from '../../bindings/crypto/finite-field.js';
 import { Poseidon } from '../../bindings/crypto/poseidon.js';
-import {
-  Group,
-  PublicKey,
-  Scalar,
-  PrivateKey,
-} from '../../provable/curve-bigint.js';
-import { Field } from '../../provable/field-bigint.js';
+import { Group, PublicKey, Scalar, PrivateKey } from './curve-bigint.js';
+import { Field } from './field-bigint.js';
 import { Nullifier } from './types.js';
 
 export { createNullifier };
