+++ src/mina-signer/src/sign-zkapp-command.ts
@@ -1,15 +1,11 @@
-import { Bool, Field, Sign, UInt32 } from '../../provable/field-bigint.js';
-import { PrivateKey, PublicKey } from '../../provable/curve-bigint.js';
+import { Bool, Field, Sign, UInt32 } from './field-bigint.js';
+import { PrivateKey, PublicKey } from './curve-bigint.js';
 import {
   Json,
   AccountUpdate,
   ZkappCommand,
 } from '../../bindings/mina-transaction/gen/transaction-bigint.js';
-import {
-  hashWithPrefix,
-  packToFields,
-  prefixes,
-} from '../../provable/poseidon-bigint.js';
+import { hashWithPrefix, packToFields, prefixes } from './poseidon-bigint.js';
 import { Memo } from './memo.js';
 import {
   Signature,
