+++ src/mina-signer/src/signature.ts
@@ -1,12 +1,12 @@
 import { blake2b } from 'blakejs';
-import { Field } from '../../provable/field-bigint.js';
+import { Field } from './field-bigint.js';
 import {
   Group,
   Scalar,
   PrivateKey,
   versionNumbers,
   PublicKey,
-} from '../../provable/curve-bigint.js';
+} from './curve-bigint.js';
 import {
   HashInput,
   hashWithPrefix,
@@ -17,7 +17,7 @@ import {
   packToFieldsLegacy,
   inputToBitsLegacy,
   HashLegacy,
-} from '../../provable/poseidon-bigint.js';
+} from './poseidon-bigint.js';
 import {
   bitsToBytes,
   bytesToBits,
